added UV sensor
this sketch needs testing.  As of right now, the "temperature sensors" are just analog values.  They need to be coupled with known resistor values, and then math has to happen to convert those values to temperatures.  Same for the UV sensor.  It's just coded to be returning UV girth units right now.  Altitude should be good.  Needs testing, though.